createNewTainted flag for the AddSecureMarksToTaintedString method (#114)
* createNewTainted flag for the AddSecureMarksToTaintedString method

* Update test/js/secure-marks.spec.js


* lint

---------

+++ src/tainted/range.h
@@ -11,7 +11,7 @@
 
 namespace iast {
 namespace tainted {
-using secure_marks_t = uint16_t;
+using secure_marks_t = uint32_t;
 class Range {
  public:
     explicit Range(int start, int end, InputInfo *inputInfo, secure_marks_t secureMarks);
